Add keymap to create temporary file
Create and open a temporary file inside `~/tmp/`.
Randomize the filename with shell command `mktemp`.

To act as a scratchpad. Jot down temporary ideas or working steps.

Discard if not required to keep. i.e. `:q!`. If need to keep, save
as with `:sav <filenametokeep>`. The file will be saved as a new
file inside vim working directory.