fix: Remove rarely use mini.session
Resort to native vim way of managing session.

To save a session, :mks .session.vim.
To load a sesion, :so .session.vim.

.session.vim is just a filename. It can be any name. For my personal
workflow, make it a dot file and with extension of .vim.

Can save multiple sessions within the directory. Useful when need
to work on different matters but same code base.

This session file is git-ignored.